+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## [Unreleased]
 
--   Resources configuration added that allows describing cpu and memory resources required in k8s by pods 
+-   Resources configuration added that allows describing cpu and memory resources required in k8s by pods
+-   Shared persistent volume can be made optional
 
 ## [0.2.0] - 2021-04-01