+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 -   Resources configuration added that allows describing cpu and memory resources required in k8s by pods 
+-   Support of S3 as DAG destination
 
 ## [0.2.0] - 2021-04-01