+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -   Resources configuration added that allows describing cpu and memory resources required in k8s by pods
 -   Shared persistent volume can be made optional
+-   Pod startup timeout is configurable, with default to 600 seconds
 
 ## [0.2.0] - 2021-04-01